fix(feature): #3104 hide closed accounts & hide help button
