+++ src/extension/features/general/hide-help/index.tsx
@@ -1,0 +1,60 @@
+import * as React from 'react';
+import { componentAppend } from 'toolkit/extension/utils/react';
+import { Feature } from 'toolkit/extension/features/feature';
+import { getToolkitStorageKey, setToolkitStorageKey } from 'toolkit/extension/utils/toolkit';
+import { HideHelpButton } from './HideHelpButton';
+import { serviceLookup } from 'toolkit/extension/utils/ember';
+
+export class HideHelp extends Feature {
+  injectCSS() {
+    return require('./index.css');
+  }
+
+  shouldInvoke() {
+    return (
+      $('#tk-hide-help').length === 0 &&
+      !!serviceLookup<YNABModalService>('modal')?.isModalOpen &&
+      !!document.querySelector('.ynab-new-settings-menu')
+    );
+  }
+
+  insertHideHelp(element: Element) {
+    if ($('#tk-hide-help', element).length) {
+      return;
+    }
+
+    componentAppend(
+      <HideHelpButton toggleHiddenState={this.setHiddenState} />,
+      element.getElementsByClassName('modal-list')[0]
+    );
+  }
+
+  invoke() {
+    const initialState = getToolkitStorageKey('hide-help', true);
+    this.setHiddenState(initialState);
+    const modalElement = document.querySelector('.ynab-new-settings-menu');
+    if (modalElement) this.insertHideHelp(modalElement);
+  }
+
+  destroy() {
+    $('#tk-hide-help').remove();
+    $('body').removeClass('toolkit-hide-help');
+  }
+
+  observe(nodes: Set<string>) {
+    if (!this.shouldInvoke()) return;
+
+    if (nodes.has('modal-overlay active  ynab-u ynab-new-settings-menu')) {
+      this.invoke();
+    }
+  }
+
+  setHiddenState = (state: boolean) => {
+    setToolkitStorageKey('hide-help', state);
+    if (state) {
+      $('body').addClass('toolkit-hide-help');
+    } else {
+      $('body').removeClass('toolkit-hide-help');
+    }
+  };
+}
