+++ src/types/ynab/services/YNABModalService.d.ts
@@ -1,0 +1,4 @@
+interface YNABModalService {
+  isModalOpen?: boolean;
+  closeModal?: (e?: unknown) => void;
+}
